feat(msg_header) Add message header including the table name (#7)
Adds a message headers in the form of <MSG_TYPE>|<HEADER>|<JSON>
A header is a string any message specific function can define (for mutations we add the table name).
The goal is to be able to allow consumers to extract some meta data regarding messages and make sense of them without having to parse the entire json, which can be a performance issue if most messages are not interesting to the consumer.

The format is text based since the plugin right now is expected to produce plain text output. We could turn everything to binary if we wanted.

Added a test for this specific case.