feat(cdc): Build the x86-64 binary and publish it as a github artifact (#10)
This builds the binary for the x86-64 platform and publishes it so that we can distribute it when generating a release.

It only supports x64 but more commonly use architectures (like i386, mips, s390 etc. :D ) can be added to the same github action as long as the postgres image for those exists.